requirements.txt
playwright>=1.35.0
weasyprint>=60.0
urllib3>=2.0.0
python-slugify>=8.0.0
